move upload code to sensible helper functions
